refactor(storage): 비슷한 조건을 함수로 묶어 수정

+++ src/storage/index.ts
@@ -1,22 +1,23 @@
 import { StorageKey } from "./types";
 
 export default class Storage {
+  private static isWindowAvailable() {
+    return typeof window !== "undefined";
+  }
+
   static getItem(key: StorageKey) {
-    return typeof window !== "undefined" ? localStorage.getItem(key) : null;
+    return this.isWindowAvailable() ? localStorage.getItem(key) : null;
   }
 
   static setItem(key: StorageKey, value: string) {
-    if (typeof window === "undefined") return;
-    localStorage.setItem(key, value);
+    if (this.isWindowAvailable()) localStorage.setItem(key, value);
   }
 
   static delItem(key: StorageKey) {
-    if (typeof window === "undefined") return;
-    localStorage.removeItem(key);
+    if (this.isWindowAvailable()) localStorage.removeItem(key);
   }
 
   static clear() {
-    if (typeof window === "undefined") return;
-    localStorage.clear();
+    if (this.isWindowAvailable()) localStorage.clear();
   }
 }
